Split Table utils.js file
